move DownloadListItem to separate item
